Add a little more function to the simple travel booking sample and simplify the command line launcher contribution configuration.
git-svn-id: https://svn.apache.org/repos/asf/tuscany/sandbox/travelsample@775672 13f79535-47bb-0310-9956-ffa450edef68

+++ contributions/introducing-tuscanyscatours-contribution/src/main/java/com/tuscanyscatours/Checkout.java
@@ -18,9 +18,11 @@
  */
 package com.tuscanyscatours;
 
+import java.math.BigDecimal;
+
 import org.osoa.sca.annotations.Remotable;
 
 @Remotable
 public interface Checkout {
-    void makePayment(double amount, String cardInfo);
+    void makePayment(BigDecimal amount, String cardInfo);
 }
